organzied the stylings a bit
