+++ farm-animal/src/components/navbar/info/Card.js
@@ -1,0 +1,161 @@
+import React from 'react'
+
+function Card() {
+  return (
+    <div className='info-row'>
+    <div className="info-column">
+        <div className='info-card'>
+
+        {/* <img src={require('/src/assets/app-info/App_Info.jpg')}/> */}
+
+        <div className='info-container'>
+            <h2>Shermol</h2>
+
+            <br></br>
+
+            <p className='card-title'>
+            Useless Team Member
+            </p>
+
+            <br></br>
+
+            <p className='card-body'>
+            Working on Info Page
+            </p>
+
+            <br></br>
+
+            <p className='card-email'>
+            SP@gmail.com
+            </p>
+
+            <br></br>
+            {/* sherrmol comment */}
+        </div>  
+        </div>
+    </div>
+
+    <div className="info-column">
+        <div className='info-card'>
+        {/* <img src="/w3images/team1.jpg" alt="Jane" style="width:100%"> */}
+        <div className='info-container'>
+            <h2>Chloe</h2>
+
+            <br></br>
+
+            <p className='card-title'>
+            Da Project Owner
+            </p>
+
+            <br></br>
+
+            <p className='card-body'>
+            Working on Shanky HW
+            </p>
+
+            <br></br>
+
+            <p className='card-email'>
+            CB@gmail.com
+            </p>
+
+            <br></br>
+
+        </div>  
+        </div>
+    </div>
+
+    <div className="info-column">
+        <div className='info-card'>
+        {/* <img src="/w3images/team1.jpg" alt="Jane" style="width:100%"> */}
+        <div className='info-container'>
+            <h2>Annie</h2>
+
+            <br></br>
+
+            <p className='card-title'>
+            *insert frog noise*
+            </p>
+
+            <br></br>
+
+            <p className='card-body'>
+            Working on Home Page
+            </p>
+
+            <br></br>
+
+            <p className='card-email'>
+            AM@gmail.com
+            </p>
+
+            <br></br>
+
+        </div>  
+        </div>
+    </div>
+
+    <div className="info-column">
+        <div className='info-card'>
+        {/* <img src="/w3images/team1.jpg" alt="Jane" style="width:100%"> */}
+        <div className='info-container'>
+            <h2>Daniel</h2>
+
+            <br></br>
+
+            <p className='card-title'>
+            Sour Patch Kid
+            </p>
+
+            <br></br>
+
+            <p className='card-body'>
+            Working on Backend
+            </p>
+
+            <br></br>
+
+            <p className='card-email'>
+            DL@gmail.com
+            </p>
+
+            <br></br>
+
+        </div>  
+        </div>
+    </div>
+
+    <div className="info-column">
+        <div className='info-card'>
+        {/* <img src="/w3images/team1.jpg" alt="Jane" style="width:100%"> */}
+        <div className='info-container'>
+            <h2>Liza</h2>
+
+            <br></br>
+
+            <p className='card-title'>
+            Team Ace
+            </p>
+
+            <br></br>
+
+            <p className='card-body'>
+            Working on Leaderboard
+            </p>
+
+            <br></br>
+
+            <p className='card-email'>
+            Liza@gmail.com
+            </p>
+            
+            <br></br>
+
+        </div>  
+        </div>
+    </div>
+    </div>
+  );
+}
+
+export default Card;
